+++ frontend/src/components/forms/queued-elements/QueuedFeaturedInstrument.mjs
