+++ frontend/src/components/forms/queued-elements/QueuedMovement.mjs
@@ -1,7 +1,7 @@
 // Import dependencies
 import React from 'react';
 import { HiOutlineMinusCircle } from 'react-icons/hi';
-import { numberToRoman } from '../../modules/utilities.mjs';
+import { numberToRoman } from '../../../modules/utilities.mjs';
 
 function QueuedMovement ({ movement, onRemove }) {
     return (
